Convert example for using events macros to Gradle
We have to inject a build for SVUnit, since that project doesn't yet provide one.

It's not a good idea to keep the unit tests at the root level. If we were to specify this directory as the test source set root, SVUnit would look starting from here for unit tests. It would go down to the `build` directory and find its link which points again to this directory and get stuck in a loop.